PT-3431: Update format of stored/serialized pedigrees
Refactored DynamicGraph:
 - ability to load pedigrees with layout from formats other than internal
 - support for new stored/external format
 - refactored code (moved heuristics into a separate file, more code reuse)
